Re-flatten remap-functions/src
