+++ lib/remap-functions/src/encode_base64.rs
@@ -1,4 +1,4 @@
-use super::Charset;
+use crate::util::Base64Charset;
 use remap::prelude::*;
 use std::str::FromStr;
 
@@ -72,7 +72,7 @@ impl Expression for EncodeBase64Fn {
                     .and_then(|v| Value::try_bytes(v).map_err(Into::into))
             })
             .transpose()?
-            .map(|c| Charset::from_str(&String::from_utf8_lossy(&c)))
+            .map(|c| Base64Charset::from_str(&String::from_utf8_lossy(&c)))
             .transpose()?
             .unwrap_or_default();
