+++ lib/remap-functions/src/lib.rs
@@ -4,18 +4,20 @@ mod util;
 mod append;
 #[cfg(feature = "assert")]
 mod assert;
-#[cfg(any(feature = "decode_base64", feature = "encode_base64"))]
-mod base64_encoding;
 #[cfg(feature = "ceil")]
 mod ceil;
 #[cfg(feature = "compact")]
 mod compact;
 #[cfg(feature = "contains")]
 mod contains;
+#[cfg(feature = "decode_base64")]
+mod decode_base64;
 #[cfg(feature = "del")]
 mod del;
 #[cfg(feature = "downcase")]
 mod downcase;
+#[cfg(feature = "encode_base64")]
+mod encode_base64;
 #[cfg(feature = "encode_json")]
 mod encode_json;
 #[cfg(feature = "ends_with")]
@@ -143,20 +145,20 @@ pub use crate::sha1::Sha1;
 pub use append::Append;
 #[cfg(feature = "assert")]
 pub use assert::Assert;
-#[cfg(feature = "decode_base64")]
-pub use base64_encoding::DecodeBase64;
-#[cfg(feature = "encode_base64")]
-pub use base64_encoding::EncodeBase64;
 #[cfg(feature = "ceil")]
 pub use ceil::Ceil;
 #[cfg(feature = "compact")]
 pub use compact::Compact;
 #[cfg(feature = "contains")]
 pub use contains::Contains;
+#[cfg(feature = "decode_base64")]
+pub use decode_base64::DecodeBase64;
 #[cfg(feature = "del")]
 pub use del::Del;
 #[cfg(feature = "downcase")]
 pub use downcase::Downcase;
+#[cfg(feature = "encode_base64")]
+pub use encode_base64::EncodeBase64;
 #[cfg(feature = "encode_json")]
 pub use encode_json::EncodeJson;
 #[cfg(feature = "ends_with")]
